requirements.txt
streamlit
pandas
numpy
scipy
statsmodels
pingouin
matplotlib
seaborn
python-dotenv
openai
python-docx
openpyxl